[PimcoreBundle] gives the Pimcore Bundle more responsability and some more default-service, therefore some changes happended to class and interface namespaces. A BC Layer is there and will be removed with final release of CoreShop. The Pimcore-Bundle is now responsible of loading Pimcore Admin Resources -> this makes the bundle more flexible and reusable for other projects.
