+++ src/CoreShop/Bundle/OrderBundle/Workflow/OrderHistoryLogger.php
@@ -14,7 +14,7 @@ namespace CoreShop\Bundle\OrderBundle\Workflow;
 
 use CoreShop\Component\Order\Model\OrderInterface;
 use CoreShop\Component\Order\Repository\OrderRepositoryInterface;
-use CoreShop\Component\Resource\Pimcore\DataObjectNoteService;
+use CoreShop\Component\Pimcore\DataObject\NoteServiceInterface;
 use Symfony\Component\Translation\TranslatorInterface;
 
 final class OrderHistoryLogger
@@ -25,7 +25,7 @@ final class OrderHistoryLogger
     private $orderRepository;
 
     /**
-     * @var DataObjectNoteService
+     * @var NoteServiceInterface
      */
     private $noteService;
 
@@ -41,13 +41,13 @@ final class OrderHistoryLogger
 
     /**
      * @param OrderRepositoryInterface $orderRepository
-     * @param DataObjectNoteService    $noteService
+     * @param NoteService    $noteService
      * @param TranslatorInterface      $translator
      * @param string                   $noteIdentifier
      */
     public function __construct(
         OrderRepositoryInterface $orderRepository,
-        DataObjectNoteService $noteService,
+        NoteServiceInterface $noteService,
         TranslatorInterface $translator,
         $noteIdentifier
     ) {
