+++ src/CoreShop/Bundle/PimcoreBundle/Resources/config/services/twig.yml
@@ -17,6 +17,6 @@ services:
     coreshop.twig.extension.pimcore_object_link_generator:
         class: CoreShop\Component\Pimcore\Twig\Extension\ObjectLinkGeneratorExtension
         arguments:
-            - '@coreshop.pimcore.object_link_generator'
+            - '@coreshop.object_link_generator'
         tags:
             - { name: twig.extension }
